whatsapp_bot/requirements.txt
# ═══════════════════════════════════════════════════════════
# WhatsApp Bot - Portal Energético MME
# Dependencias del proyecto
# ═══════════════════════════════════════════════════════════

# ────────────────────────────────────────────────────────────
# Framework Web
# ────────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ────────────────────────────────────────────────────────────
# WhatsApp Integration
# ────────────────────────────────────────────────────────────
twilio==8.11.1
python-multipart==0.0.6

# ────────────────────────────────────────────────────────────
# Telegram Bot Integration - 100% GRATIS
# ────────────────────────────────────────────────────────────
python-telegram-bot==20.7

# ────────────────────────────────────────────────────────────
# Database & Cache
# ────────────────────────────────────────────────────────────
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.1.4

# ────────────────────────────────────────────────────────────
# Task Queue
# ────────────────────────────────────────────────────────────
celery==5.3.6
celery[redis]==5.3.6

# ────────────────────────────────────────────────────────────
# AI & NLP
# ────────────────────────────────────────────────────────────
openai==1.12.0
spacy==3.7.2
langchain==0.1.6
langchain-community==0.0.20

# ────────────────────────────────────────────────────────────
# Charts & Visualization
# ────────────────────────────────────────────────────────────
plotly==5.18.0
matplotlib==3.8.2
kaleido==0.2.1
pillow==10.2.0

# ────────────────────────────────────────────────────────────
# Web Scraping & Browser Automation
# ────────────────────────────────────────────────────────────
playwright==1.41.1
selenium==4.17.2

# ────────────────────────────────────────────────────────────
# Cloud Storage
# ────────────────────────────────────────────────────────────
boto3==1.34.34
minio==7.2.3

# ────────────────────────────────────────────────────────────
# HTTP Clients
# ────────────────────────────────────────────────────────────
httpx==0.26.0
requests==2.31.0

# ────────────────────────────────────────────────────────────
# Utilities
# ────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pytz==2024.1

# ────────────────────────────────────────────────────────────
# Security
# ────────────────────────────────────────────────────────────
cryptography==42.0.2

# ────────────────────────────────────────────────────────────
# Monitoring
# ────────────────────────────────────────────────────────────
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# ────────────────────────────────────────────────────────────
# Testing
# ────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
